+++ views/default/resources/elgg-news/custom_list_view.php
@@ -20,7 +20,8 @@ if ($featured) {
         [ 
             'name' => 'featured',
             'value' => NewsOptions::NEWS_YES, 
-            'operand' => '=' ],
+            'operand' => '=', 
+        ],
     ];
 }